stats_service/requirements.txt
grpcio==1.62.0
grpcio-tools==1.62.0
clickhouse-driver==0.2.6
kafka-python==2.0.2
